TASK: refactor EditorEnvelope
