+++ packages/neos-ui/src/Containers/RightSideBar/Inspector/InspectorEditorEnvelope/index.spec.js
